+++ spec/migrations/table_definition_spec/001_add_status_to_quux.rb
@@ -1,0 +1,7 @@
+class AddStatusToQuux < ActiveRecord::Migration::Current
+  def change
+    create_table :quux do |t|
+      t.enum :status, as: "status_type"
+    end
+  end
+end